Refactoring: - Separated views and templates into category folders. - Extracted logic from views. - Small fixes to schedule scripts. - Usage improvement to ajax script.
