--- templates/courses-design.html
@@ -1,84 +1,0 @@
-<html dir="rtl">
-	<head>
-		<!-- Latest compiled and minified CSS -->
-		<link 
-		  rel="stylesheet"
-		  href="https://cdn.rtlcss.com/bootstrap/v4.2.1/css/bootstrap.min.css"
-		  integrity="sha384-vus3nQHTD+5mpDiZ4rkEPlnkcyTP+49BhJ4wJeJunw06ZAp+wzzeBPUXr42fi8If"
-		  crossorigin="anonymous">
-
-		<!-- Latest compiled and minified JavaScript -->
-		<script
-		  src="https://cdn.rtlcss.com/bootstrap/v4.2.1/js/bootstrap.min.js"
-		  integrity="sha384-a9xOd0rz8w0J8zqj1qJic7GPFfyMfoiuDjC9rqXlVOcGO/dmRqzMn34gZYDTel8k"
-		  crossorigin="anonymous"></script>
-		
-		<link rel="stylesheet" href="../static/styles/style.css" />
-        <meta charset="utf-8">
-
-
-	</head>
-	<body>
-		<nav class="navbar navbar-expand-lg navbar-light bg-navbar">
-		  <a class="navbar-brand" href="#">Skedj</a>
-		  <button class="navbar-toggler" type="button" data-toggle="collapse" data-target="#navbarSupportedContent" aria-controls="navbarSupportedContent" aria-expanded="false" aria-label="Toggle navigation">
-			<span class="navbar-toggler-icon"></span>
-		  </button>
-
-		  <div class="collapse navbar-collapse" id="navbarSupportedContent">
-			<ul class="navbar-nav mr-auto">
-			  <li class="nav-item active">
-				<a class="nav-link" href="#"></a>
-			  </li>
-			  <li class="nav-item">
-				<a class="nav-link" href="#">קורסים</a>
-			  </li>
-			  <li class="nav-item">
-				<a class="nav-link" href="#">בניית מערכת</a>
-			  </li>
-                </li>
-			  <li class="nav-item">
-				<a class="nav-link" href="#">אודות</a>
-			  </li>
-			</ul>
-		  </div>
-		</nav>
-		
-		<section>
-
-			<!--Grid row-->
-			<div class="row">
-
-				<!--Grid column-->
-				<div class="col-md-12">
-					<div class="card search-bg card-image"
-                         style="background-image: url(../static/images/search-bg.jpeg);">
-
-					</div>
-				</div>
-				<!--Grid column-->
-
-			</div>
-			<!--Grid row-->
-
-		</section>
-		<!--Projects section v.4-->
-
-        <div class="container">
-            <div class="row">
-                <div class="col-md-4 course-trailer py-5 px-5">
-                    <div class="course-trailer-content">
-                        course 1
-                    </div>
-                </div>
-                <div class="col-md-4 course-trailer py-5 px-5">
-                  One of three columns
-                </div>
-                <div class="col-md-4 course-trailer py-5 px-5">
-                  One of three columns
-                </div>
-            </div>
-        </div>
-		
-	</body>
-</html>
