+++ templates/courses/course-item.html
