+++ templates/courses/courses.html
@@ -5,7 +5,7 @@
 {% block head %}
     <script src={% static 'scripts/handlebars.js' %}></script>
     <script name="course-single" type="text/x-handlebars-template">
-        {% raw_include 'handlebars/course-item.html' %}
+        {% raw_include 'courses/course-item.html' %}
     </script>
     <script>
         function renderSearchResults(courses) {
