+++ templates/schedule/schedule-course-autocomplete-item.html
