+++ templates/schedule/schedule-course-item.html
