+++ templates/schedule/schedule.html
@@ -6,19 +6,19 @@
 
     <script src={% static 'scripts/handlebars.js' %}></script>
     <script name="schedule-course-autocomplete-item" type="text/x-handlebars-template">
-        {% raw_include 'handlebars/schedule-course-autocomplete-item.html' %}
+        {% raw_include 'schedule/schedule-course-autocomplete-item.html' %}
     </script>
     <script name="schedule-course-item" type="text/x-handlebars-template">
-        {% raw_include 'handlebars/schedule-course-item.html' %}
+        {% raw_include 'schedule/schedule-course-item.html' %}
     </script>
     <script name="schedule-course-item-group-header" type="text/x-handlebars-template">
-        {% raw_include 'handlebars/schedule-course-item-group-header.html' %}
+        {% raw_include 'schedule/schedule-course-item-group-header.html' %}
     </script>
     <script name="schedule-course-item-group-item" type="text/x-handlebars-template">
-        {% raw_include 'handlebars/schedule-course-item-group-item.html' %}
+        {% raw_include 'schedule/schedule-course-item-group-item.html' %}
     </script>
     <script name="schedule-single-event" type="text/x-handlebars-template">
-        {% raw_include 'handlebars/schedule-single-event.html' %}
+        {% raw_include 'schedule/schedule-single-event.html' %}
     </script>
     <script>document.getElementsByTagName("html")[0].className += " js";</script>
 {% endblock head %}
@@ -57,16 +57,18 @@
             <div class="col-lg-9">
                 <nav>
                     <div class="nav nav-tabs" id="nav-tab">
-                        <a class="nav-item nav-link active" id="semester-1-tab" data-toggle="tab" href="#semester-1-wrapper">סמסטר א'</a>
-                        <a class="nav-item nav-link" id="semester-2-tab" data-toggle="tab" href="#semester-2-wrapper">סמסטר ב'</a>
+                        <a class="nav-item nav-link active" id="semester-1-tab" data-toggle="tab"
+                           href="#semester-1-wrapper">סמסטר א'</a>
+                        <a class="nav-item nav-link" id="semester-2-tab" data-toggle="tab" href="#semester-2-wrapper">סמסטר
+                            ב'</a>
                     </div>
                 </nav>
                 <div class="tab-content mt-3" id="nav-tabContent">
                     <div class="tab-pane fade show active" id="semester-1-wrapper" data-schedule="0">
-                        {% include 'schedule-timetable.html' with semester=1 %}
+                        {% include 'schedule/schedule-timetable.html' with semester=1 %}
                     </div>
-                    <div class="tab-pane fade" id="semester-2-wrapper"  data-schedule="1">
-                        {% include 'schedule-timetable.html' with semester=2 %}
+                    <div class="tab-pane fade" id="semester-2-wrapper" data-schedule="1">
+                        {% include 'schedule/schedule-timetable.html' with semester=2 %}
                     </div>
                 </div>
             </div>
@@ -74,5 +76,10 @@
         <div class="clear-fix"></div>
     </div>
     <script src="{% static 'scripts/schedule-util.js' %}" type="text/javascript"></script>
-    <script src="{% static 'scripts/schedule.js' %}" type="text/javascript"></script>
+    <script src="{% static 'scripts/schedule-helper.js' %}" type="text/javascript"></script>
+    <script src="{% static 'scripts/schedule-logic.js' %}" type="text/javascript"></script>
+    <script>
+        CSRF_TOKEN = '{{ csrf_token }}';
+        initWithChoices({{ choices|safe }});
+    </script>
 {% endblock content_body %}
